#38: refactoring the element data and rendering life-cycle
* renamed files
* moved files
* minor refactoring
