+++ package.json
@@ -21,7 +21,8 @@
         "vite::dev": "vite",
         "vite::preview": "vite preview",
         "vitest": "vitest run",
-        "vitest::watch": "vitest watch"
+        "vitest::debug": "vitest watch",
+        "vitest::dev": "NODE_ENV=production HEADLESS=true vitest watch --mode production "
     },
     "private": true,
     "dependencies": {
