+++ src/components/complex/coding/coding.ts
@@ -1,6 +1,6 @@
 import {html, TemplateResult} from 'lit'
 import {customElement}        from 'lit/decorators.js'
-import {BaseElement}          from '../../../internal/base'
+import {BaseElement}          from '../../../internal'
 import {DisplayConfig}        from '../../../types'
 import {CodingData}           from './coding.data'
