+++ src/components/complex/signature/signature.ts
@@ -1,6 +1,6 @@
 import {html, TemplateResult} from 'lit'
 import {customElement}        from 'lit/decorators.js'
-import {BaseElement}          from '../../../internal/base'
+import {BaseElement}          from '../../../internal'
 import {DisplayConfig}        from '../../../types'
 
 import {SignatureData} from '../../foundation/bundle/bundle.data'
