+++ src/components/resources/substance/substance-ingredient.backbone.ts
@@ -1,6 +1,6 @@
 import {html, TemplateResult}                                         from 'lit'
 import {customElement}                                                from 'lit/decorators.js'
-import {BaseElement}                                                  from '../../../internal/base'
+import {BaseElement}                                                  from '../../../internal'
 import {DisplayConfig}                                                from '../../../types'
 import {SubstanceIngredientData, SubstanceIngredientReferenceData}    from './substance-ingredient.data'
 import {isSubstanceIngredientConcept, isSubstanceIngredientReference} from './substance-ingredient.type-guard'
