+++ src/internal/BaseElement.ts
@@ -1,11 +1,11 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import {PropertyValues, TemplateResult}    from 'lit'
-import {DisplayConfig}                     from '../../types'
-import {FhirElementData, ValidationErrors} from './data/fhir-data-element.data'
+import {DisplayConfig}                     from '../types'
+import {FhirElementData, ValidationErrors} from './base/data/fhir-data-element.data'
 
-import {Decorated}              from './Decorated'
-import {EmptyResult}            from './presentable'
-import {FhirPresentableElement} from './presentable/fhir-presentable-element'
+import {Decorated}              from './base/Decorated'
+import {EmptyResult}            from './base/presentable'
+import {FhirPresentableElement} from './base/presentable/fhir-presentable-element'
 
 /**
  * Represents a base element in the FHIR data model. This is the class to extend when creating components.
