+++ src/internal/base/index.ts
@@ -4,7 +4,6 @@ export * from './DataHandling'
 export * from './Decorated'
 export * from './Rendering'
 export * from './Templating'
-export * from './base-element'
 export * from './data/'
-export * from './oneOf'
+export * from './util/oneOf'
 export * from './presentable/'
