+++ src/internal/resource/Resource.ts
@@ -2,7 +2,8 @@ import {html, PropertyValues, TemplateResult} from 'lit'
 import {property}                             from 'lit/decorators.js'
 import {PrimitiveType}                        from '../../components/primitive/type-converters/type-converters'
 import {DisplayConfig, DisplayMode}           from '../../types'
-import {BaseElement, Decorated, NoDataSet}    from '../base'
+import {Decorated, NoDataSet}                 from '../base'
+import {BaseElement}                          from '../BaseElement'
 import {ContextProviderController}            from '../contexts'
 import {ResourceData}                         from './domain-resource.data'
